Implement simple locking primitives
These are spinlocks. Add locking to malloc, console and FSP code. These
code path will be used with multiple processors up, especially while in
the "runtime" environment so let's have all the infrastructure ready.

We move abort() to the core as well, and hook it up with a simple
backtrace (no symbols yet !)

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>