Fix fputs() arguments and add missing fputc()
As odd as it sounds, fputs really wants the string first and
the stream second. Not doing it right is a problem since gcc
will silently replace fprintf() calls with fputs() calls when
there is only a single string argument.

It will also replace one-argument fputs() with fputc(), so change
putc to fputc() and implement putc() as a macro, as the man page
says it can be.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>