Base relocation support
We build with -fpie and process relocations early at boot.

This only supports RELA/R_PPC64_RELATIVE type relocations, so we
have to tweak the assembly a bit to avoid generating different onces,
the advantage however is that these relocs do not rely on the previous
content of the relocated code or data, so they can be applied multiple
times easily.

We will still have to backup our data segment if we implement "rebooting"
a copy of ourselves for runtime though.

To simplify things for the asm we change a bit how the boot stack is
defined, and we now use a fixed offset in head.S for the boot CPU
semaphore.

We still don't actually attempt to run at a different address yet.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>