Some minor SPIRA updates
We should have the act_len field set in processor init. Also
add a comment based on my observation that those values are
ignored by the FSP which seems to always boot from 0x180 with
an MSR value of 0xf000000000000000.

Also clear the NACA bit indicating that we support PCIA, we don't

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>