Implements support for FSP interrupts
With this patch, FSP interrupts work, though the kernel doesn't currently
make much use of them, it does make mbox operations snappier.

Additionally, they are implement in such a way that calls to
opal_poll_events() deep down a callchain should hopefully not "eat"
interrupts. IE udbg or kernel console doing a bunch of poll's will
still cause an interrupt to eventually occur later on, so one can
now reliably rely on interrupts on various state bit changes (*)
even when mixing in opal_poll_events() calls along the way.

(*) Console out doesn't generate interrupts, this is an FSP limitation
unfortunately.

We will be able to eventually make good use of that in the kernel by
replacing poll loops in various places such as real time clock or
nvram commands, but only after the original OPAL either goes away or
gets fixed to behave the same way.

In addition this adds a file that will eventually host GX related
bits and pieces (we need to configure a GX register for the PSI interrupt
to work properly).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>