Add documentation of the OPAL API changes
Sapphire changes the behaviour of some OPAL APIs in ways that diverge
either from the current internal specification when it exists or from
OPALv2 implementation. In some cases, the behaviour is unspecified
in the spec and the OPAL v2 implementation acts as a reference.

Those changes have been made to increase the API usefulness and in such
a way that is backward compatible with existing Linux Kernel
implementations. IE. We have carefully designed the changes to work with
the way current Linux exploits the OPAL APIs.

This document those changes and will be used as a reference for subsequent
changes to Linux to exploit those new functionalities.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>