p7ioc: Handle mangling of XIVE for extended server numbers
To address limitations in the size of the server number, our systems
are setup to hijack the 3 top priority bits to extend the server number.

We need to apply this mangling to our P7IOC ICS code or we'll end up
sending interrupts to the wrong targets which can result in checkstops.

However, we also need to keep a priority of "ff" as-is since the HW
handles that specific value as indicating "masked". Thus we lose the
server number when masking an interrupt.

Since Linux does read-modify-write sequences on the xive, I've added
a cache so we can return the last value programmed.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>