memory: create our own device tree
Instead of having our own address_range structure, use device nodes.

As an intermediate step, we create our own root and flatten that when
asked, but this will eventually be a real in-mem device tree and we
won't flatten it ourselves.

We actually don't need to parse all the way down to the DIMMs; the ram
areas are enough, so we cut out some code here.

Signed-off-by: Rusty Russell <rusty@au.ibm.com>