cpus: create our own device tree
We generate in-memory device tree nodes for the cpus as we discover
them, rather than post-generating them.  We still keep the
cpu_threads[] array, since that's accessed via asm and also has
entries for threads (only primary threads aka cpus get a device node).

We create our own root for the moment, but this will eventually be
the global device-tree.

Signed-off-by: Rusty Russell <rusty@au.ibm.com>