Move opal and interrupt nodes into in-memory device tree.
We rename add_interrupt_controller() to add_interrupt_controllers() and
add all of them to the tree at once, not just the @0 node.

The opal nodes are related, but don't belong under hdata/: we need to
add them ourselves whatever the source is for the device tree.  This
these are added at the end, not in parse_machine().

Signed-off-by: Rusty Russell <rusty@au.ibm.com>