Big interrupt rework
Interrupt sources now register ranges of ISNs with the core
interrupt layer. A source must provide get_xive() and set_xive()
callbacks.

It can also optionally provide an interrupt() callback in which
case it is an interrupt meant to be handled by us on behalf of the
OS, and it will be added to the "opal-interrupts" property.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>