pci: EEH API updates
Add a "severity" enum which provides a clearer indication of the
scope of an error (and it's recoverability). Add a new variant
of eeh_freeze_status() that returns this information.

Add a next_error() call which returns the severity of the next
pending error on a PHB along with the first PE affected (in
frozen state)

Define an event indicating that there is a pending error on
a given PHB

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>