CPU thread/stacks rework
This finishes cutting the ties between hdata/cpu.c and core/cpu.c

The CPU stacks are now in a fixed memory location after SKIBOOT_BASE
and the cpu_thread structures are at the begining of each stack which
makes it a lot easier to get to them.

The cpu_thread's are initialized from core/cpu.c purely based on
the device-tree content.

The P7/P7+/P8 differences are now entirely handled at runtime

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>