OPAL: Don't spin in skiboot for opal_pci_reset()
Acccording to Ben's suggestion: It's not safe to spin in OPAL API
since the OS would disable interrupt while calling into OPAL API.
The patch refactors opal_pci_reset() so that we won't spin in
skiboot anymore. The cost for that is the extra fields introduced
to PHB (struct phb) to remember the current type and stage of
required reset. On the other hand, the OS has to keep calling the
OPAL API with same parameters until getting failure or success.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>