P7IOC: Apply hardware workaround to hard reset
The hard set is being done through SHPC. At the mean while, we also
might hot reset the root bridge. The hardware workaround, which is
to disable PCIe link training and enable that after hot resetting
the root bridge, isn't applied. The patch applied the hardware
workaround according to Ben's suggestion.

Apart from that, the procedure for hard (fundamental) reset is changed
for a bit. We would force hot reset on the root bridge instead of
doing that on detecting broken PCIe link.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>