PCI: Doing fundamental reset after power-on
Before doing PCI probe, we assure that power is on for the specific
PHB (slot). After that, we need do fundamental reset instead of
hot reset as Ben suggested.

It's notable that the change would be applied to PHB complete reset
as well. That means, the PHB will be forced with fundamental reset
when taking complete reset.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>