P7IOC: Remove duplicate states
When we're powering on the PHB, fundamental reset would be issued
to that after power is on. At the same time, we have state machine
to cover request of fundamental reset as well. For those 2 cases,
we needn't separate SM (State Machine) states. The patch removes
those duplicated states as Ben suggested.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>