Implement machine check handler
Currently, we do nothing on exceptions, which is probably not what the
OS intends when registering OPAL exceptions handlers.

This change implements a handler for the machine check interrupt: we
decode values from SRR0, SRR1, DSISR and DAR, then convert that to a
OPAL machine check event. We then pass this to the kernel, with MSR.ME
enabled.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Rusty Russell <rusty@au.ibm.com>