Rework stack frame setup in exception path
Currently, the stack frame setup is a bit off; we declare an ABI gap,
but that goes *below* the stack pointer (ie, negative offset from r1).
Also, we store data in the ABI gap, in exactly the place that a callee
is supposed to save LR. This means that the restored LR will be
incorrect.

This change restructures struct stack_frame to look more like a ppc64
stack frame. We declare the standard fields, then an area for
frame-local data, with a new asm offset definition, STACK_LOCALS.

This fixes the LR issue, as we save and restore from an area that is
safe from modification by the callee. As a bonus, the stack frame size
is reduced.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Rusty Russell <rusty@au.ibm.com>