HDAT/chiptod: Handle FSP-cleared chiptod data
Apparently the FSP may clear-out the chiptod data from HDAT; in this
case, we don't add the chiptod nodes to the device tree. This means that
the chiptod driver needs to handle missing chiptod nodes, and pick a
suitable primary.

This change moves the missing-HDAT-handing code to the the hdat parser;
if we find a zero-ed out chiptod entry in hdat, we use the same logic to
pick a primary.

This means that we get the same DT for FSP vs. non-FSP entry, and we
don't need to special-case the chiptod driver.

Signed-off-by: Jeremy Kerr <jeremy.kerr@au.ibm.com>