fsp_console: Fix console on Firebird blades
We need to register DVS way earlier since we get the open
messages during the OPL. If we aren't registered at that
point we miss them and don't setup the console properly.

Also do some hacks to deal with both DVS and blademm hitting
the console. Ultimately we want to sort that out more cleanly
and put some exclusion there, possibly removing the support
for partition !=0 for actual OS console.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>