Rework/cleanup chip numbering
The HDAT provides a "proc_chip_id" which we used here or there (especially
on P7) as our "chip id". However it doesn't really relate to the HW chip
numbers (it's a pHyp thing), and our code made various incorrect
assumptions about it.

Additionally, we were manipulating on P7 a chip_id here or there which
was made of 3 node bits and 2 chip bits, while converting sometimes to
a "GCID" with a T bit inserted. This was also source of much confusion.

This does a big overhaul. Now, the chip ID everywhere in skiboot and in
the device tree is the "XSCOM" chip ID which corresponds to the fabric
chip ID.

On P8, this is 3 bits of node and 3 bits of chip. On P7, this is 3 bits of
node, a T bit (always 0 for now) and 2 bits of chip. This is consistent
across the board.

Since this doesn't always match the top of the PIR (it does on P8 but not
on P7 since the PIR doesn't have the T bit), the old xscom_pir_to_gcid()
function is replaced by a new pir_to_chip_id() which will do the right
conversion always.

All the information related to chip numbering are also moved to a pair
of dedicated files, chip.c and chip.h

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>