PHB3: Improve link detection
Some IBM machines seem to have a presence detect pin set even when
nothing is plugged (riser cards ?). We work around this two ways:

  - Use the new proprietary presence detect "A/B" pins when told to
do so by the device-tree (which we only do on HDAT boots for now)
to detect the b'11 combination which means "nothing connected".

  - Break the link up state machine into two phases: A 1s phase waiting
for the electrical link to establish followed by a 9s phase waiting for
the link nego to complete. This allows an early bailout if the electricals
don't come up which should hopefully be quick.

Note: We still need to reduce that 9s second phase and implement retries
at lower speeds to support broken old v1 devices

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>