op-panel: Avoid allocations in oppanel output
We currently try to minimise allocations in op_display by re-using a
struct fsp message. However, this still means we need to allocate the
first time we call op_display; if we die in the allocator init path, we
can end up in infinite recursion.

Instead, just use a static message.

Signed-off-by: Jeremy Kerr <jeremy.kerr@au.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>