PHB3: Setup valid default PE#s in the RTT at IODA reset time
Linux starts probing before it maps PEs. Under some circumstances
an error can occur during the probe done by Linux. The current setting
of having all PE# set as invalid can cause the HW to fail more
completely (PHB fence) or the simics simulator to abort if that
happened.

As a workaround, we set the RTT to 0x0000 instead of 0xffff during
IODA reset. This has the side effect of matching the content of
the pe_remap in Linux at probe time, so all accesses are considered
as PE#0 until otherwise decided.

Note that an explicit unmap PE operation still uses 0xffff since by
the time Linux might do this (it doesn't today), PE#0 is valid and
assigned to some device and should not be hijacked. When we support
map/unmap (hotplug/SR-IOV) in Linux we will have to make sure we
don't do probes on non-mapped RIDs.

The failing scenario prompting that change is the latest simics
trying to apparently generate an error when Linux accesses the root
complex cfg space register 0x10 which isn't implemented in sim (it
appears to be in HW as it should be, just returns 0's).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>