fsp-rtc: Cache values read from RTC
We're seeing that the FSP may take upwards of 10 seconds to read the
RTC, particularly when the RTC is sourced from NTP.

This change introduces a RTC cache. Duing skiboot init, we perform one
read to get an initial RTC value, and record the timebase. When the OS
requests a read during runtime, we send a read request to the FSP. If
this read request times out (currently 1.5 sec), we return a value
calculated from the cache, plus a timebase-derived offset.

When a timed-out read does return, we use the RTC value to update our
cache.

Signed-off-by: Jeremy Kerr <jeremy.kerr@au.ibm.com>