hdata/test: add unit tests for hdata parsing.
We have some canned spira and hdata heap dumps: we can create a fairly
simple test framework which turns it into a device tree using the real
code.

Currently it just dumps out the tree in a very dumb way: we just make
sure it doesn't assert() or abort().  The next step is to compare with
real device trees.

Signed-off-by: Rusty Russell <rusty@au.ibm.com>