lpc/uart: Rework LPC access mechanisms and how we expose the UART
This adds a "cycle type" argument to LPC accessors.

Additionally, the LPC device node is fixed up (until HB is updated) and
we also add the UART under control of a config option.

Finally a new ENABLE_DUMMY_CONSOLE option allows to control whether the
UART is exposed directly to Linux or via the OPAL console mechanism.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>