fsp-console: Change default to 0 (network/IPMI)
This changes the default console passed to the loader kernel to 0
(network DVS or IPMI console, ie hvc0 in Linux).

You can change the setting by changing the "i5/OS partition boot mode"
in the FSP "Power On/Off System" menu. The default "A" corresponds
to the network / IPMI console (hvc0). "B" is serial port 1 and "C" is
serial port 2.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>