p7ioc: Fix initialization
Some machines hang during initialization. It appears that we need
some delay between the various MMIO stores when doing the
inits. A read back seems to do the trick so let's do that
unconditionally. The spec actually mentions that this is what
pHyp does.

In addition, change the RGC init procedure to match more closely
the documentation, and remove the setting of a bunch of HSS registers
which we lave to whatever value was set by the FSP.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>