psi: Clean up extra interrupts numbering
Use definitions within the PSI "block", so that one can do:

    get_psi_interrupt(chip) + P8_IRQ_PSI_<foo>

Where <foo> is FSP, OCC, FSI, ...

Also use the indices in interrupts.h for the configuration and
the selection of interrupt numbers in the PSI code instead of
hard coded number, which means it is now possible to *change*
the numbers. Thus is one of the interrupts is meant to be part
of the block passed to Linux, the numbers can be shuffled
(the number emitted by a given HW source is SW configured).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>