P7IOC: Support opal_pci_phb_mmio_enable()
The OPAL API is used to enable or disable one of 16 M64 BARs. It's
notable that we should have configured the base/size for that M64
BAR correctly before enabling it. Otherwise, it might incur strange
issues like fenced AIB. As usual, we save the M64 BARs into the
corresponding cache so that we can restore them after PHB complete
reset.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>