P7IOC: Support opal_pci_set_phb_mem_window()
The OPAL API is used to set the base address and mask for specified
M64 BAR. It's notable that the parameter has been changed for a bit.
We need designate the whole size of that M64 BAR instead of M64
segment size. Since the OPAL API is shared by multiple chips and PHB3
possiblly have PE-owned M64 BAR (exclusively). On the other hand,
to designating the size of whole BAR instead of segment will give the
OS more flexibility.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>