PHB3: PCI-CFG access blocked or by ASB
Similar to the case of P7IOC: we need blocking the PCI-CFG access
while we're doing reset to avoid further PCIE errors. Also, the
AIB channel isn't reliable for PCI-CFG access when the PHB is fenced,
so we have to use ASB in that case.

The patch also adjust the format of macros PHB3_PCI_CFG_READ and
PHB3_PCI_CFG_WRITE, but the logic (except blocking or access via
ASB) isn't affected.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>