phb3: Cached fenced PHB state
For each cycle of PCI-CFG access, we're checking PHB fenced state
via NFIR, which is retrieved through xscom. The xscom path is a
bit slow as Ben pointed. Also as Ben suggested, we cache the PHB
fenced status and use the catched status for check, instead of that
one from NFIR register. It would improve the performance for a bit.

Signed-off-by: Gavin Shan <shangw@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>