PSI: Fix interrupt bit mappings for PSI host bridge interrupts
There is a discrepancy between P7 Book4 and SCOM docs about which bits
are the PSI/FSP interrupts in the PSI Host Bridge Control register.
Based on BenH and the FSP team's inputs and from tests on hardware,
the SCOM docs are right. Fix code appropriately.

This really doesn't affect P8. Tested on P7.

Signed-off-by: Ananth N Mavinakayanahalli <ananth@in.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>