elog: Various cleanups and fixes
Things that should not be done without the lock like dereferencing
a log entry element or changing the OPAL state (because some other
CPU might just be finishing a fetch and trying to set it), that
sort of things. Also implement fsp_elog_set_head_state() as I
actually envisioned it.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>