fsp: Fix queue hang on completion of message with a response
If we have more than one message queue in a command class and the
head message is waiting for a reply, then a response to it fails
to trigger a re-check of the queue and thus a send of the next
message in that command class, which is thus delayed until some
other message completion triggers that recheck.

This fixes it by returning from fsp_handle_incoming() whether a
message in a queue was completed as a result of the incoming message
and that result is then used to trigger a recheck of the queues.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>