psi: Don't enable PSI TCEs unless there is an FSP
Enable TCEs has the side effect of limiting cronus/ecmd "getmemdma" to
the area mapped by TCEs. Due to issues with the ecmd bindings for Perl
"getmemproc" doesn't work properly from the BML scripts and "getmemdma"
is the only way to get to the internal console.

So on FSP-less machines, let's keep that working.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>